ai-agent/ui/requirements.txt
anthropic==0.39.0
python-telegram-bot==20.7
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
python-dotenv==1.0.0
python-multipart

# RAG / Document processing
sentence-transformers
numpy
pypdf
python-docx
tiktoken
